backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Flash Facts Fun Backend API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "minio>=7.2.0",
    "python-multipart>=0.0.12",
    "pillow>=11.0.0",
    "requests>=2.32.0",
]
